requirements.txt
 Pillow>=8.0.0
numpy>=1.19.0
PyQt5>=5.15.0
mss>=6.0.0 